+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 > 👉 Version numbers here refer to the Docker image for the demo app. For the libraries history, see the README of each library in this workspace.
 
+- 2026-03-02: improve `CitSchemeService.toString`.
 - 2026-02-22:
   - updated Angular and packages.
   - fixes to geo location editor.